Replace Enterprise store API dialog with banner
